Styling and validation cleanup
