+++ frontend/src/components/settings/tabs/GeneralSettings/AllowedWebsitesList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Input, Switch, Button, Space, Tag, Tooltip } from "antd";
+import { Input, Switch, Button, Flex, Tag, Tooltip } from "antd";
 import { InfoCircleOutlined } from "@ant-design/icons";
 import { Plus } from "lucide-react";
 
@@ -37,14 +37,14 @@ const AllowedWebsitesList: React.FC<AllowedWebsitesListProps> = ({
   };
 
   return (
-    <div className="space-y-4">
-      <div className="flex items-center justify-between">
-        <span className="flex items-center gap-2">
+    <Flex vertical gap="small">
+      <Flex align="center" justify="space-between" wrap gap="small">
+        <Flex align="center" justify="start" gap="small">
           Allowed Websites List
           <Tooltip title="When enabled, Magentic-UI will only be able to visit websites you add to the list below.">
             <InfoCircleOutlined className="text-secondary hover:text-primary cursor-help" />
           </Tooltip>
-        </span>
+        </Flex>
         {cachedWebsites.length === 0 && (
           <Switch
             checked={allowedlistEnabled}
@@ -57,28 +57,27 @@ const AllowedWebsitesList: React.FC<AllowedWebsitesListProps> = ({
                 handleUpdateConfig({ allowed_websites: undefined });
               }
             }}
+            style={{flex: "none"}}
           />
         )}
-      </div>
-      <Space direction="vertical" style={{ width: "100%" }}>
-        {allowedlistEnabled || cachedWebsites.length > 0 ? (
-          <>
-            <div className="flex w-full gap-2">
+      </Flex>
+      <Flex vertical gap="large">
+        {(allowedlistEnabled || cachedWebsites.length > 0) &&
+          <Flex vertical gap="small">
+            <Flex gap="large">
               <Input
                 placeholder="https://example.com"
                 value={websiteInput}
                 onChange={(e) => setWebsiteInput(e.target.value)}
                 onPressEnter={addWebsite}
-                className="flex-1"
+                // className="flex-1"
               />
               <Button icon={<Plus size={16} />} onClick={addWebsite}>
                 Add
               </Button>
-            </div>
-            <div>
-              {cachedWebsites.length === 0 ? (
-                <div></div>
-              ) : (
+            </Flex>
+            <Flex wrap="wrap" gap="large">
+              {(cachedWebsites.length > 0) && 
                 cachedWebsites.map((site, index) => (
                   <Tag
                     key={index}
@@ -89,14 +88,12 @@ const AllowedWebsitesList: React.FC<AllowedWebsitesListProps> = ({
                     {site}
                   </Tag>
                 ))
-              )}
-            </div>
-          </>
-        ) : (
-          <div className="text-secondary italic"></div>
-        )}
-      </Space>
-    </div>
+              }
+            </Flex>
+          </Flex>
+        }
+      </Flex>
+    </Flex>
   );
 };
