+++ frontend/src/styles/global.css
@@ -396,7 +396,8 @@ a:hover {
 }
 
 .ant-collapse-header {
-  @apply bg-secondary text-primary border-primary !important;
+  @apply bg-primary text-primary border-primary !important;
+  align-items: center !important;
 }
 
 .ant-collapse-content {
@@ -445,4 +446,8 @@ div#___gatsby,
 div#gatsby-focus-wrapper {
   height: 100%;
   /* border: 1px solid green; */
+}
+
+.ant-input {
+  @apply bg-primary !important;
 }
